Ultimos get y delete

+++ Models/Empleado.cs
@@ -5,7 +5,7 @@
         public string Name { get; set; }
         public string LastName { get; set; }
         public string Document { get; set; }
-        public int Salary { get; set; }
+        public double Salary { get; set; }
         public char Gender { get; set; }
         public string Position { get; set; }
         public DateOnly StartDate { get; set; }
